Decouple Mdoc Zk engine from ZK registry
