+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/HolderAgent.kt
@@ -25,7 +25,7 @@ import at.asitplus.wallet.lib.jws.SignJwt
 import at.asitplus.wallet.lib.jws.SignJwtFun
 import at.asitplus.wallet.lib.procedures.dcql.DCQLQueryAdapter
 import at.asitplus.wallet.lib.procedures.iso.DeviceRetrievalProcedure
-import at.asitplus.wallet.lib.zk.iso.IsoMdocZkBackendRegistry
+import at.asitplus.wallet.lib.zk.iso.IsoMdocZkEngine
 import io.github.aakira.napier.Napier
 import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
@@ -46,12 +46,13 @@ class HolderAgent @JvmOverloads constructor(
     private val signVerifiablePresentation: SignJwtFun<VerifiablePresentationJws> =
         SignJwt(keyMaterial, JwsHeaderCertOrJwk()),
     private val signKeyBinding: SignJwtFun<KeyBindingJws> = SignJwt(keyMaterial, JwsHeaderNone()),
+    private val mdocZkEngine: IsoMdocZkEngine = IsoMdocZkEngine.Default,
     private val verifiablePresentationFactory: VerifiablePresentationFactory =
         VerifiablePresentationFactory(
             keyMaterial = keyMaterial,
             signVerifiablePresentation = signVerifiablePresentation,
             signKeyBinding = signKeyBinding,
-            mdocZkBackendRegistry = IsoMdocZkBackendRegistry.Default
+            mdocZkEngine = mdocZkEngine
         ),
     private val difInputEvaluator: PresentationExchangeInputEvaluator = PresentationExchangeInputEvaluator,
 ) : Holder {
