+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/VerifiablePresentationFactorySdJwtTest.kt
@@ -8,6 +8,7 @@ import at.asitplus.testballoon.matrix.fixture
 import at.asitplus.testballoon.matrix.matrixSuite
 import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.data.rfc3986.toUri
+import at.asitplus.wallet.lib.zk.iso.IsoMdocZkEngine
 import at.asitplus.wallet.lib.zk.iso.IsoMdocZkBackendRegistry
 import com.benasher44.uuid.uuid4
 import io.kotest.matchers.maps.shouldHaveSize
@@ -66,7 +67,7 @@ val VerifiablePresentationFactorySdJwtTest by matrixSuite {
             object {
                 val verifiablePresentationFactory = VerifiablePresentationFactory(
                     keyMaterial = holderKeyMaterial,
-                    mdocZkBackendRegistry = IsoMdocZkBackendRegistry()
+                    mdocZkEngine = IsoMdocZkEngine(IsoMdocZkBackendRegistry())
                 )
                 val sdJwtCredential = sdJwtCredential
             }
